pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "graphocr"
version = "0.1.0"
description = "Hybrid Graph-OCR Pipeline: Deterministic Trust Layer for Insurance Claims"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "python-multipart>=0.0.17",

    # OCR
    "paddleocr>=2.9.0",
    "paddlepaddle>=3.0.0",
    "surya-ocr>=0.8.0",

    # LLM / Agent
    "langgraph>=0.2.60",
    "langchain>=0.3.12",
    "langchain-community>=0.3.12",
    "langchain-openai>=0.3.0",  # vLLM uses OpenAI-compatible API
    "openai>=1.58.0",           # vLLM client

    # DSPy
    "dspy>=2.5.0",

    # Knowledge Graph
    "neo4j>=5.26.0",

    # RAG — Vector Store + Embeddings
    "chromadb>=0.5.0",
    "sentence-transformers>=3.3.0",

    # Monitoring
    "langsmith>=0.2.10",

    # Task Queue
    "redis>=5.2.0",
    "arq>=0.26.1",

    # Storage
    "minio>=7.2.12",

    # Image Processing
    "Pillow>=10.2.0",
    "pdf2image>=1.17.0",
    "opencv-python-headless>=4.10.0",
    "numpy>=1.26.0",

    # Language Detection
    "fasttext-wheel>=0.9.2",

    # Utilities
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "uuid7>=0.1.0",
    "typer>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "testcontainers[neo4j]>=4.9.0",
    "ipykernel>=6.29.0",
]

[project.scripts]
graphocr = "graphocr.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/graphocr"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (no external services needed)",
    "integration: Integration tests (may require external services)",
    "layer1: Layer 1 — OCR + Spatial Foundation (token, assembler, reading order, failure classifier)",
    "layer2: Layer 2 — Adversarial Verification (rule engine, self-healing, knowledge graph)",
    "layer3: Layer 3 — Tiered Inference (traffic controller, circuit breaker)",
    "rag: RAG temporal policy retrieval tests",
    "dspy: DSPy metrics and optimization tests",
    "smoke: Fast critical-path subset for quick validation",
]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
